fix compile warnings

+++ include/CPU.hpp
@@ -43,7 +43,7 @@ class CPU
         std::string parseName();
         void CPUUpdate(std::vector<CPUCore>& cores);
 
-        int numCores = 0;
+        size_t numCores = 0;
 
     private:
         struct UptimeData
